# api.yaml
openapi: 3.0.3
info:
  title: Placement Portal API
  version: 1.0.0
  description: OpenAPI definition for the JSON APIs exposed by the Placement Portal Flask app.
servers:
  - url: http://localhost:5000
    description: Local development server
...
paths:
...
  /api/students:
    get:
      tags: [Students]
      summary: List students
      operationId: listStudents
      responses:
        '200':
          description: Student list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentSummary'
    post:
      tags: [Students]
      summary: Create a student and its user account
      operationId: createStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentRequest'
      responses:
        '201':
          description: Student created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateStudentResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/students/{student_id}:
    put:
      tags: [Students]
      summary: Update a student
      operationId: updateStudent
      parameters:
        - $ref: '#/components/parameters/StudentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStudentRequest'
      responses:
        '200':
          description: Student updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Students]
      summary: Delete a student and its user account
      operationId: deleteStudent
      parameters:
        - $ref: '#/components/parameters/StudentId'
      responses:
        '200':
          description: Student deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
...
    StudentId:
      name: student_id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    StudentSummary:
      type: object
      properties:
        student_id:
          type: integer
        name:
          type: string
      required: [student_id, name]
    CreateStudentRequest:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        roll_no:
          type: string
      required: [name, username, password, roll_no]
    UpdateStudentRequest:
      type: object
      properties:
        name:
          type: string
        roll_no:
          type: string
        phone:
          type: string
        department:
          type: string
        degree:
          type: string
        batch_year:
          type: integer
        cgpa:
          type: number
          format: float
        resume_url:
          type: string
    CreateStudentResponse:
      type: object
      properties:
        message:
          type: string
        student_id:
          type: integer
        user_id:
          type: integer
      required: [message, student_id, user_id]
...